Se agrega popup de confirmacion al momento de eliminar usuarios

+++ frontend-application/src/app/adm-usuarios/usuarios/editar-usuario/editar-usuario.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { Usuario } from '../../../model/usuario.model';
 import { UsuarioService } from './../../../services/usuario.service';
 import { Router } from '@angular/router';
+import {MatDialog, MatDialogRef} from '@angular/material';
+import { ConfirmacionPopupComponent } from '../../confirmacion-popup/confirmacion-popup.component';
+
 
 @Component({
   selector: 'app-editar-usuario',
@@ -11,7 +14,8 @@ import { Router } from '@angular/router';
 export class EditarUsuarioComponent implements OnInit {
 
   constructor(private _router: Router,
-              private _usuarioService: UsuarioService) { }
+              private _usuarioService: UsuarioService,
+              private dialog: MatDialog) { }
 
   usuarios: Usuario[];
   usuarioSelected: Usuario = null;
@@ -29,16 +33,26 @@ export class EditarUsuarioComponent implements OnInit {
   }
 
   deleteUser(usuario: Usuario){
-    this._usuarioService.deleteUser(usuario).subscribe(
-      res => {
-        console.log(res);
-        alert("Se elimino el usuario correctamente!");
-        this._router.navigate(['/home/gestion/usuarios']);
-      },
-      error => {
-        console.log(error);
-      }
-    );
+    const dialogRef = this.dialog.open(ConfirmacionPopupComponent,{
+      width: '40%',
+      data: { mensaje: "Desea eliminar el usuario?"}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+
+        if (result && result == "true"){
+          this._usuarioService.deleteUser(usuario).subscribe(
+            res => {
+              console.log(res);
+              alert("Se elimino el usuario correctamente!");
+              this._router.navigate(['/home/gestion/usuarios']);
+            },
+            error => {
+              console.log(error);
+            }
+          );
+        }
+    });
   }
 
   modificarUser(usuario: Usuario){
